add function: rest student password

+++ src/main/java/Review/r_server/userInfo/UserInfoController.java
@@ -4,20 +4,16 @@ import Review.r_basic.r_b_user.User;
 import Review.r_basic.r_b_user.UserService;
 import Review.r_server.paperInfo.PaperInfo;
 import Review.r_server.paperInfo.PaperInfoService;
+import Review.r_util.MD5Util;
 import Review.r_util.R_FileUtils;
-import com.github.pagehelper.PageHelper;
-import com.github.pagehelper.PageInfo;
 import org.apache.commons.io.FileUtils;
 import org.apache.log4j.Logger;
-import org.apache.poi.ss.formula.functions.Mode;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.support.ManagedMap;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
-
-import javax.servlet.ServletException;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.*;
@@ -45,28 +41,44 @@ public class UserInfoController {
     @Autowired
     private PaperInfoService paperInfoService;
 
+    /**
+     * 查看个人信息
+     *
+     * @param model
+     * @param request
+     * @return
+     */
     @RequestMapping(value = "/info", method = RequestMethod.GET)
-    public String userInfo(Model model) {
+    public String userInfo(Model model, HttpServletRequest request) {
+        UserInfo userInfo = (UserInfo) request.getSession().getAttribute("userInfo");
         model.addAttribute("ifCheckStudent", 0);
-        logger.info("user:\t" + 1 + "check_info");
+        logger.info("user:" + userInfo.getReal_name() + "check personal info ");
         return "functionJsp/userInfo/userInfo";
     }
 
+    /**
+     * 更新个人信息
+     * @param studentInfo
+     * @param request
+     * @return
+     */
     @RequestMapping(value = "/info", method = RequestMethod.POST)
     public String infoUpdate(UserInfo studentInfo, HttpServletRequest request) {
         UserInfo userInfo = (UserInfo) request.getSession().getAttribute("userInfo");
-        logger.info("user:\t" + 1 + "check_info");
-        System.out.println("studentInfo:::::::::::" + studentInfo);
-        System.out.println("userInfo:::::::::::" + userInfo);
-        userInfo.setPhone(studentInfo.getPhone() == null && studentInfo.getPhone().equals("") ? userInfo.getPhone() : studentInfo.getPhone());
-        userInfo.setEmail(studentInfo.getEmail() == null && studentInfo.getEmail().equals("") ? userInfo.getEmail() : studentInfo.getEmail());
+        userInfo.setPhone(studentInfo.getPhone() == null || studentInfo.getPhone().equals("") ? userInfo.getPhone() : studentInfo.getPhone());
+        userInfo.setEmail(studentInfo.getEmail() == null || studentInfo.getEmail().equals("") ? userInfo.getEmail() : studentInfo.getEmail());
         userInfoService.updateUserInfo(userInfo);
         request.getSession().removeAttribute("userInfo");
         request.getSession().setAttribute("userInfo", userInfo);
+        logger.info("user:\t" + userInfo.getReal_name() + "update personal info ");
         return "redirect:info";
     }
 
-    //跳转到学生信息添加页面
+    /**
+     * 跳转到学生信息添加页面
+     * @param request
+     * @return
+     */
     @RequestMapping(value = "/addStudent", method = RequestMethod.GET)
     public String addStudent(HttpServletRequest request) {
         UserInfo userInfo = (UserInfo) request.getSession().getAttribute("userInfo");
@@ -78,7 +90,7 @@ public class UserInfoController {
     @ResponseBody
     public Map<String, String> addStudentPost(String username, String password, String place, String stu_tch_name, String real_name, HttpServletRequest request) {
         UserInfo userInfo = (UserInfo) request.getSession().getAttribute("userInfo");
-        System.out.println(username + " " + password + " " + real_name + " " + stu_tch_name + " " + place);
+        logger.info(username + " " + password + " " + real_name + " " + stu_tch_name + " " + place);
         Map<String, String> msg = new ManagedMap<String, String>();
         User newUser = new User();
         newUser.setUsername(username);
@@ -99,16 +111,16 @@ public class UserInfoController {
             if (userInfoService.insertUserInfo(newUserInfo) != 0) {
                 //给上传的用户添加 角色
                 userService.insertUserRole(newUserInfo.getUser_id(), (long) 5);
-                System.out.println("^^^^^^^^^ user " + newUserInfo.getReal_name() + " success!");
+                logger.info("^^^^^^^^^ user " + newUserInfo.getReal_name() + " success!");
                 msg.put("result", "1");
                 msg.put("tip", "添加成功");
             } else {
-                System.out.println("^^^^^^^^^ user " + newUserInfo.getReal_name() + " insertUserInfo error!");
+                logger.info("^^^^^^^^^ user " + newUserInfo.getReal_name() + " insertUserInfo error!");
                 msg.put("result", "0");
                 msg.put("tip", "学号冲突");
             }
         } else {
-            System.out.println("^^^^^^^^^ user " + newUserInfo.getReal_name() + " insertUser error!");
+            logger.info("^^^^^^^^^ user " + newUserInfo.getReal_name() + " insertUser error!");
             msg.put("result", "0");
             msg.put("tip", "学号冲突");
         }
@@ -117,11 +129,16 @@ public class UserInfoController {
     }
 
 
-    //跳转到 学生列表
+    /**
+     *跳转到 学生列表
+     * @param model
+     * @param request
+     * @return
+     */
     @RequestMapping(value = "/studentList", method = RequestMethod.GET)
     public String studentList(Model model, HttpServletRequest request) {
         UserInfo userInfo = (UserInfo) request.getSession().getAttribute("userInfo");
-        logger.info("!!!!!!!!!!!!!!!!!!!!!!!!");
+        logger.info("!@# " + userInfo.getReal_name() + " turn to studentList");
 
         List<UserInfo> studentList = userInfoService.selectStudentList();
         for (UserInfo info : studentList) {
@@ -135,7 +152,11 @@ public class UserInfoController {
     }
 
 
-    //跳转到学生信息导入页面
+    /**
+     *
+     * 跳转到学生信息导入页面
+     * @return
+     */
     @RequestMapping(value = "/importStudentListPage", method = RequestMethod.GET)
     public String importStudentListPage() {
         return "functionJsp/userInfo/importStudentListPage";
@@ -182,9 +203,7 @@ public class UserInfoController {
                     .getRealPath("/serverFile/import_data") + "\\"
                     + "学生信息表[" + new SimpleDateFormat("yyyy-MM-dd").format(new Date()) + "].xlsx";
             File f = new File(path); // 输入要删除的文件位置
-            if (f.exists()) {
-                f.delete();
-            }
+            if (f.exists()) f.delete();
             FileUtils.copyInputStreamToFile(import_student_file.getInputStream(), f);
             Map<String, Object> importResult = userInfoService.importStudentFromExcel(path);
             if ("success".equals(importResult.get("result"))) {
@@ -193,7 +212,6 @@ public class UserInfoController {
                 List<UserInfo> errorUserInfoList = new ArrayList<UserInfo>();
                 User user = new User();
 
-                System.out.println("userInfoList@@@@@@@@@@@@@@" + userInfoList.size());
                 for (int i = 0; i < userInfoList.size(); i++) {
                     UserInfo info = userInfoList.get(i);
 
@@ -208,34 +226,36 @@ public class UserInfoController {
                         if (userInfoService.insertUserInfo(info) != 0) {
                             //给上传的用户添加 角色
                             userService.insertUserRole(info.getUser_id(), (long) 5);
-                            System.out.println(i + "^^^^^^^^^ user " + info.getReal_name() + " success!");
+                            logger.info(i + "^^^^^^^^^ user " + info.getReal_name() + " insertUserInfo success!");
                         } else {
-                            System.out.println(i + "^^^^^^^^^ user " + info.getReal_name() + " insertUserInfo error!");
+                            logger.info(i + "^^^^^^^^^ user " + info.getReal_name() + " insertUserInfo error!");
                             errorUserInfoList.add(info);
                         }
                     } else {
-                        System.out.println(i + "^^^^^^^^^ user " + info.getReal_name() + " insertUser error!");
+                        logger.info(i + "^^^^^^^^^ user " + info.getReal_name() + " insertUser error!");
                         errorUserInfoList.add(info);
                     }
                 }
-                System.out.println("errorUserInfoList##############" + errorUserInfoList.size());
+                logger.info("errorUserInfoList##############" + errorUserInfoList.size());
                 msg.put("errorUserInfoList", errorUserInfoList);
                 msg.put("result", "1");
                 msg.put("tip", "导入成功");
-
             } else {
                 msg.put("result", "0");
                 msg.put("tip", "导入失败");
             }
-
-
-            msg.put("result", "1");
-            msg.put("tip", "导入成功");
         }
         return msg;
     }
 
 
+    /**
+     * 显示学生信息
+     * @param stu_username
+     * @param request
+     * @param model
+     * @return
+     */
     @RequestMapping(value = "/{stu_username}/info", method = RequestMethod.GET)
     public String studentInfo(@PathVariable String stu_username, HttpServletRequest request, Model model) {
         UserInfo userInfo = (UserInfo) request.getSession().getAttribute("userInfo");
@@ -243,27 +263,57 @@ public class UserInfoController {
         UserInfo studentInfo = userInfoService.selectUserInfoByUserId(student.getId());
         model.addAttribute("studentInfo", studentInfo);
         model.addAttribute("ifCheckStudent", 1);
-        logger.info("user:\t" + 1 + "check_info");
+        logger.info("user:" + userInfo.getReal_name() + " check " + studentInfo.getReal_name() + " info");
         return "functionJsp/userInfo/userInfo";
     }
 
+    /**
+     * 更新学生信息
+     * @param studentInfo
+     * @param stu_username
+     * @param request
+     * @return
+     */
     @RequestMapping(value = "/{stu_username}/info", method = RequestMethod.POST)
     public String studentInfoUpdate(UserInfo studentInfo, @PathVariable String stu_username, HttpServletRequest request) {
-//        UserInfo userInfo = (UserInfo) request.getSession().getAttribute("userInfo");
-        logger.info("user:\t" + 1 + "check_info");
-        System.out.println(stu_username);
-        System.out.println("studentInfo::::" + studentInfo.toStuString());
+        UserInfo userInfo = (UserInfo) request.getSession().getAttribute("userInfo");
         User student = userService.selectByUsername(stu_username);
-        System.out.println("student::::" + student);
         UserInfo studentInfoOld = userInfoService.selectUserInfoByUserId(student.getId());
-        System.out.println("studentInfoOld::::" + studentInfoOld.toStuString());
         studentInfoOld.setReal_name(studentInfo.getReal_name() == null || studentInfo.getReal_name().equals("") ? studentInfoOld.getReal_name() : studentInfo.getReal_name());
         studentInfoOld.setStu_tch_name(studentInfo.getStu_tch_name() == null || studentInfo.getStu_tch_name().equals("") ? studentInfoOld.getStu_tch_name() : studentInfo.getStu_tch_name());
         studentInfoOld.setPlace(studentInfo.getPlace() == null || studentInfo.getPlace().equals("") ? studentInfoOld.getPlace() : studentInfo.getPlace());
         studentInfoOld.setEmail(studentInfo.getEmail() == null || studentInfo.getEmail().equals("") ? studentInfoOld.getEmail() : studentInfo.getEmail());
         studentInfoOld.setPhone(studentInfo.getPhone() == null || studentInfo.getPhone().equals("") ? studentInfoOld.getPhone() : studentInfo.getPhone());
-        System.out.println("studentInfoOld::::" + studentInfoOld.toStuString());
         userInfoService.updateUserInfo(studentInfoOld);
+        logger.info("user:" + userInfo.getReal_name() + " update " + studentInfo.getReal_name() + " info");
         return "redirect:info";
+    }
+
+    /**
+     * 重置学生密码
+     *
+     * @param username
+     * @param request
+     * @return
+     */
+    @RequestMapping(value = "/restPassword", method = RequestMethod.POST)
+    @ResponseBody
+    public Map<String, Object> restPassword(String username, HttpServletRequest request) {
+        UserInfo userInfo = (UserInfo) request.getSession().getAttribute("userInfo");
+        Map<String, Object> msg = new HashMap<String, Object>();
+        logger.info("user:\t" + 1 + "check_info");
+        User student = userService.selectByUsername(username);
+        UserInfo studentInfo = userInfoService.selectUserInfoByUserId(student.getId());
+        student.setPassword(new MD5Util().compute(student.getUsername() + student.getUsername()));
+        if (userService.updateUser(student) == 1) {
+            msg.put("result", "1");
+            msg.put("tip", "重置密码成功！");
+            logger.info("user:" + userInfo.getReal_name() + " rest " + studentInfo.getReal_name() + "`s password success  !!!!");
+        } else {
+            msg.put("result", "0");
+            msg.put("tip", "重置密码失败！");
+            logger.info("user:" + userInfo.getReal_name() + " rest " + studentInfo.getReal_name() + "`s password failure  !!!!");
+        }
+        return msg;
     }
 }
